marshale v0.1.2 Marshale.ModelMagic

A collection of helper functions for Marshale.ModelMagic.defmodel/1.

Functions

convert_function(type)
(since 0.1.0)

Returns AST for a Marshale.ModelUtil.convert/2 call based on a passed in type.

To be more specific, this returns the AST for the anonymous function call, which in non-AST terms would look like this:

&Marshale.ModelUtil.convert(&1, <type>)

Examples

iex> Macro.to_string(Marshale.ModelMagic.convert_function(String))
"&(Marshale.ModelUtil.convert(&1, String))"

deconvert_function(type)
(since 0.1.2)

Returns AST for a Marshale.ModelUtil.deconvert/2 call based on a passed in type.

See Marshale.ModelMagic.convert_function/1 for what this is the reverse of.

defmodel(do_block)
(macro) (since 0.1.0)

Creates a model in the current module.

This does the following:

typedocs accordingly

types accordingly

creates a __from_map__/1 which converts to the types you specify

this only Marshale.ModelUtil.convert/2s to types where Marshale.ModelMagic.is_marshale_module?/1 returns true.

creates a __to_map__/1 which changes the data back to the external form. This is mainly for things like a datetime type or the like.

creates a struct in the current module with the given fields

document_line(line_number, simple_ast)
(since 0.1.0)

Returns AST for a @typedoc statement from a line number, and the SimpleAST™️.

This function gets the documentation for a variable using Marshale.ModelMagic.get_typedoc/2, then creates a @typedoc [documentation], which it then returns the AST for (or nil if the variable has no documentation associated with it).

Examples

iex> Macro.to_string(Marshale.ModelMagic.document_line(
...>   3,
...>   [{2, :document, "Onwards and upwards"}]
...> ))
"@typedoc(\"Onwards and upwards\")"

document_type(simple_ast_variable, simple_ast)
(since 0.1.0)

Returns a list which when Enum.flat_map/2-ed is the AST lines for documenting a variable.

Takes in the SimpleAST™️ form of a variable and the whole SimpleAST™️ region, and outputs said AST lines for documenting the variable.

Examples

iex> Macro.to_string(Marshale.ModelMagic.document_type(
...>   {3, :set, :foo, :bar},
...>   [{2, :document, "baz"},
...>    {3, :set, :foo, :bar}]
...> ))
"[@typedoc(\"baz\"), @type(foo :: bar)]"

get_typedoc(line_number, simple_ast)
(since 0.1.0)

Gets the typedoc for a variable (or nil is there is none).

Takes in the line number of the variable, and the SimpleAST™️ of the block. It returns a string which is what the variable is typedoc-ed as, or nil.

Examples

iex> Marshale.ModelMagic.get_typedoc(3, [
...>    {1, :document, "Hello, world!"},
...>    {3, :set, :foo, :Bar}
...> ])
"Hello, world!"

iex> Marshale.ModelMagic.get_typedoc(3, [{3, :set, :foo, :Bar}])
nil

handle_sigil_s(content)
(since 0.1.0)

Takes in a ~s's or ~S's AST and returns the inner string.

This may prevent interpolation from ~s (I have not checked), but is a good compromise between simplicity and feature-complete-ness. If the passed in argument is not ~s's or ~S's AST, then this function just returns said argument.

Examples

iex> Marshale.ModelMagic.handle_sigil_s(
...>   {:sigil_S, [line: 2], [{:<<>>, [line: 2], ["Aloha!"]}, []]}
...> )
"Aloha!"

iex> Marshale.ModelMagic.handle_sigil_s("Aloha!")
"Aloha!"

is_marshale_module?(module)
(since 0.1.0)

Detects whether a passed in module has __from_map__/1.

This is the only test of whether of something is a Marshale module, allowing for basic modules to be assembled by hand, allowing for type conversions not of the map to struct type.

Examples

iex> defmodule Snowflake do
...>   def __from_map__(string) when is_binary(string) do
...>     String.to_integer(string)
...>   end
...>
...>   def __from_map__(default) do
...>     default
...>   end
...> end
iex> Marshale.ModelMagic.is_marshale_module?(Snowflake)
true

is_possibly_module(module)
(since 0.1.2)

Returns whether something is possibly a module in Marshale's eyes.

All this does is check is_atom/1, or if it's a list with one component, is_atom/1 of that component. This is a less aggressive optimization, but is used here because we can't be sure that the types used are fully loaded.

is_to_mappable?(module)
(since 0.1.2)

Detects whether a module has a __to_map__/1 function.

This is for use in things that have an external representation, for example a timestamp which has to be sent in ISO8601 format.

manipulate_line(ast)
(since 0.1.0)

Turns AST to a SimpleAST™️ representation.

This helps out with processing it, as it turns the AST which has tons of encapsulation into 2 possible different flat lines. They are:

{line_number, :set, variable_name, type}

{line_number, :document, documentation}

Just do quote do: a = :b and look at the mess that is for processing easily. That's exactly why this method exists.

Examples

iex> Marshale.ModelMagic.manipulate_line(
...>   {:=, [line: 1], [
...>     {:foo, [line: 1], nil},
...>     {:__aliases__, [line: 1], [:bar]}
...>   ]}
...> )
{1, :set, :foo, :"Elixir.bar"}

iex> Marshale.ModelMagic.manipulate_line(
...>   {:@, [line: 1], [{:typedoc, [line: 1], ["Aloha!"]}]}
...> )
{1, :document, "Aloha!"}

map_update_function(field_and_function_tuple)
(since 0.1.0)

Returns AST for a Map.update/4 call based on field name and function.

The input should be a tuple of field name and function. To be specific, the output and the input function are both AST. I highly recommend using Marshale.ModelMagic.convert_function/1 to generate the input function AST.

Examples

iex> Macro.to_string(Marshale.ModelMagic.map_update_function({
...>   :foo,
...>   Marshale.ModelMagic.convert_function(String)
...> }))
"Map.update(:foo, nil, &(Marshale.ModelUtil.convert(&1, String)))"

namespace_types(type)
(since 0.1.0)

Namespaces an atom.

This both prepends :Elixir. to it, and also combines multiple names to allow things such as :"Marshale.ModelMagic" instead of just :ModelMagic and the like.

Examples

iex> function_exported?(:String, :downcase, 1)
false
iex> function_exported?(Marshale.ModelMagic.namespace_types(:String), :downcase, 1)
true

iex> Marshale.ModelMagic.namespace_types([:Marshale, :ModelMagic])
Marshale.ModelMagic

pipe(list)
(since 0.1.0)

Constructs a pipeline from a list.

In more specific, it should be able to a list of anything, but in practice, that "anything" is normally atoms. It returns the AST for a pipeline, too.

Examples

iex> Macro.to_string(Marshale.ModelMagic.pipe([100, :foo, :bar]))
"100 |> :foo |> :bar"

iex> piped = Marshale.ModelMagic.pipe([100, (quote do: div(2)), (quote do: div(5))])
{:|>, [], [{:|>, [], [100, {:div, [], [2]}]}, {:div, [], [5]}]}
iex> Macro.to_string(Macro.expand_once(piped, __ENV__))
"div(div(100, 2), 5)"

pop_tail(list)
(since 0.1.0)

Removes the last element from a list.

This function is basically taken from StackOverflow, but is here because Marshale.ModelMagic.pipe/1 requires it as a helper function.

Examples

iex> Marshale.ModelMagic.pop_tail([])
{:err, "Can't pop the tail of an empty list"}

iex> Marshale.ModelMagic.pop_tail([1, 2, 7, 3])
[1, 2, 7]

struct_to_map()
(since 0.1.2)

Creates a Struct.to_map/1 call.

Examples

iex> Macro.to_string(Marshale.ModelMagic.struct_to_map)
"Map.from_struct(this)"

tail(list)
(since 0.1.0)

Returns the last element of a list.

The code is basically Marshale.ModelMagic.pop_tail/1, but slightly modified. It's only here due to the fact Elixir includes no such utility to do this, and because both Marshale.ModelMagic.get_typedoc/2 and Marshale.ModelMagic.pipe/1 need it.

Examples

iex> Marshale.ModelMagic.tail([])
{:err, "Can't take the tail of an empty list"}

iex> Marshale.ModelMagic.tail([1, 2, 7, 3])
3