marshale v0.1.2 Marshale.ModelUtil

Functions

convert(item, type)
(since 0.1.0)

Converts an item(s) to a type.

To be specific, this only converts if the type is a "marshale type". This is determined by Marshale.ModelMagic.is_marshale_module?/1.

Examples

iex> Marshale.ModelUtil.convert(nil, Integer)
nil

iex> Marshale.ModelUtil.convert("7", [Integer])
"7"

iex> Marshale.ModelUtil.convert(["7"], [Integer])
["7"]

deconvert(item, type)
(since 0.1.2)

The reverse of Marshale.ModelUtil.convert/2.

Note that this is akin to serialization, while Marshale is mostly focused on deserialization (external data with incorrect types -> internal data with correct types). This means this is not very powerful, but it should serve its purposes to anybody who wants a simple data marshalling library that goes both ways.

Also, to aim for simplicity when writing custom Marshale modules, the required function (__to_map__/1) is not checked as part of Marshale.ModelMagic.is_marshale_module?/1.